refactor: simplify route structure and clean up components
- Remove (main) route group and move pages to root
- Consolidate layout into root layout.tsx
- Merge PropertyCard into FeaturedProperties component
- Update Sanity environment configuration
- Update CLAUDE.md with monorepo commands
