+++ apps/frontend/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "dzts-inmobiliaria",
-  "version": "0.1.0",
+  "name": "dzts-website",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
