--- apps/frontend/src/app/(main)/layout.tsx
@@ -1,22 +1,0 @@
-import BootstrapClient from "@/components/BootstrapClient";
-import Header from "@/components/Header";
-import Footer from "@/components/Footer";
-import "bootstrap-icons/font/bootstrap-icons.css";
-import "bootstrap/dist/css/bootstrap.min.css";
-import { SanityLive } from "@/sanity/lib/live";
-
-export default function MainLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <>
-      <Header />
-      <main>{children}</main>
-      <Footer />
-      <BootstrapClient />
-      <SanityLive />
-    </>
-  );
-}
