+++ apps/frontend/src/app/propiedades/[slug]/page.tsx
