+++ apps/frontend/src/app/propiedades/page.tsx
