+++ apps/frontend/src/sanity/env.ts
@@ -1,5 +1,5 @@
 export const apiVersion =
-  process.env.NEXT_PUBLIC_SANITY_API_VERSION || '2025-05-06'
+  process.env.NEXT_PUBLIC_SANITY_API_VERSION || '2026-01-23'
 
 export const dataset = assertValue(
   process.env.NEXT_PUBLIC_SANITY_DATASET,
